Reapply r260096.
Expand the simplify boolean expression check to handle implicit conversion of integral types to bool and improve the handling of implicit conversion of member pointers to bool.

Implicit conversion of member pointers are replaced with explicit comparisons to nullptr.

Implicit conversions of integral types are replaced with explicit comparisons to 0.

Patch by Richard Thomson.

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@260681 91177308-0d34-0410-b5e6-96231b3b80d8